requirements.txt
trafilatura
python-dotenv
requests
